Tests/LibWeb/Text/expected/ServiceWorker/service-worker-ready.txt

ready is a Promise: true
ready is the same object on second access: true
